refactor(ai): split anthropic tests into a sibling file under the r8 loc cap
Behavior-preserving move of the test module (all 35 tests verbatim) into
anthropic_tests.rs via the answer_assist_tests.rs precedent; anthropic.rs
drops from 1540 to 951 lines, back under the 1400 hard cap.
